refactor and take certificate ordering logic to new class
